refactor(short-conv): make cache_indices int64-canonical
Collapse ShortConvMetadata's dual index fields to a single int64
cache_indices (the canonical index dtype), hoisted once per step. Models
consume it directly with no cast: ZAYA1 cca_decode uses it for index_copy_;
LFM2 / LFM2-MoE pass it to causal_conv1d. The compiled CUDA causal_conv1d
kernel requires int32, so causal_conv1d_fn/update narrow int64->int32 right
at the kernel call site (the Triton path reads int64 natively). No .to()
cast remains in any short-conv model.

ZAYA CCA/config/mod_tp CPU suites pass; GPU confirms int64 flows through
causal_conv1d (output + state match int32).

+++ python/sglang/srt/models/lfm2.py
@@ -284,7 +284,7 @@ class Lfm2ShortConv(nn.Module):
                 self.conv_weight,
                 self.conv_bias,
                 activation=None,
-                conv_state_indices=meta.cache_indices.to(torch.int32),
+                conv_state_indices=meta.cache_indices,
             )
         else:
             # Prefill: multiple tokens, use varlen kernel
@@ -294,7 +294,7 @@ class Lfm2ShortConv(nn.Module):
                 self.conv_weight,
                 self.conv_bias,
                 query_start_loc=meta.query_start_loc,
-                cache_indices=meta.cache_indices.to(torch.int32),
+                cache_indices=meta.cache_indices,
                 has_initial_state=meta.has_initial_state,
                 conv_states=conv_state,
                 activation=None,
